requirements.txt
fastapi==0.104.1
openai>=1.12.0
psycopg2-binary==2.9.9
python-docx==1.1.0
PyPDF2==3.0.1
pydantic==2.5.0
python-multipart==0.0.6
requests==2.31.0
PyJWT>=2.8.0
